Also add to ruleConverters.ts, alphabetically
