+++ docs/Creating a Rule Converter.md
@@ -6,11 +6,11 @@ Adding a new rule converter to `tslint-to-eslint-config` is a relatively straigh
 For your convenience, a starter script is included that sets up the files:
 
 ```shell
-node ./script/new-converter --eslint output-name --tslint input-name
+node ./script/newConverter --eslint output-name --tslint input-name
 ```
 
 If the lint rule includes arguments, add the `--sameArguments` flag above to have starter code generated for that as well.
 
 ```shell
-node ./script/new-converter --eslint output-name --tslint input-name --sameArguments
+node ./script/newConverter --eslint output-name --tslint input-name --sameArguments
 ```
